Update README.md
The program is now complete and ready to execute. After running the program, the model will be saved to a file called 'tf_model.h5' and the test accuracy of the model will be printed to the console.

please tell me execution procedure I need to follow
To execute the program, you will need to have Python and TensorFlow installed on your computer. Once you have installed the necessary libraries, you will need to run the program using a Python interpreter. Depending on the platform you are using, this can be done in several different ways. For example, if you are using Windows, you can open the command line and type in 'python <program_name>.py' to execute the program. For more detailed instructions on running Python programs, please refer to the documentation for your platform.
which python library
The Python library that is used in the program is TensorFlow. TensorFlow is an open source library for numerical computation, developed by Google and used in many machine learning applications. It is available for download and installation from the official TensorFlow website.